Allow passing a hash function to as.list() and print() for keys.
Currently public keys *cannot be generated or even printed* on
FIPS-compliant systems because these systems do not permit use of the
MD5 algorithm:

    > k <- openssl::rsa_keygen(2048)
    > pub <- k$pubkey
    Error: OpenSSL error in EVP_DigestInit_ex: disabled for fips
    # Using another approach:
    > pub <- as.list(k)$pubkey
    Error: OpenSSL error in EVP_DigestInit_ex: disabled for fips
    # Not even printing will work, since it calls fingerprint(), too:
    > print(k)
    Error: OpenSSL error in EVP_DigestInit_ex: disabled for fips

This commit allows passing a new 'hashfun' argument to the as.list() and
print() methods for keys and public keys, which are in turn passed
through to the underlying call to fingerprint(). This enables the
following workarounds on FIPS systems:

    > pub <- as.list(k, hashfun = openssl::sha256)$pubkey
    > print(k, hashfun = openssl::sha256)
    [2048-bit rsa private key]
    sha256 20:c4:fc:84:44:5f:97:f4:2d:ef:15:db:b9:62:e6:e1:73:52:f9:c4:03:be:dd:f3:5c:5d:7f:8c:41:f9:ee:13

In order to preserve backwards compatibility, the hash function still
defaults to MD5, but this could probably be changed in the future.
Recent versions of OpenSSL's command-line tools, for example, have
switched to SHA-256.

Some new unit tests have been added to test these features.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>